+++ apps/backoffice-v2/src/pages/Home/components/StatsCard/StatsCard.tsx
@@ -1,3 +1,4 @@
+import { ArrowDownRightIcon, ArrowUpRightIcon } from 'lucide-react';
 import type { FunctionComponent } from 'react';
 import { Link } from 'react-router-dom';
 
@@ -5,25 +6,38 @@ import { Card } from '@/common/components/atoms/Card/Card';
 import { CardContent } from '@/common/components/atoms/Card/Card.Content';
 import { CardHeader } from '@/common/components/atoms/Card/Card.Header';
 import { ctw } from '@/common/utils/ctw/ctw';
+import { WarningSvg } from '@/common/components/atoms/icons';
+import { WarningFilledSvg } from '@ballerine/ui';
 
 type StatsCardProps = {
   prefix?: string;
   href?: string;
-  count: number;
+  value: number | string;
   title: string;
   description: string;
   className?: string;
   valueClassName?: string;
   centered?: boolean;
+  style?: React.CSSProperties;
+  alert?: boolean;
+  tendency?: {
+    value: number | string;
+    direction: 'up' | 'down';
+    kind: 'positive' | 'negative' | 'neutral';
+  };
 };
 
 export const StatsCard: FunctionComponent<StatsCardProps> = ({
   prefix = '',
-  count,
+  value,
   title,
   description,
   href,
   centered,
+  tendency,
+  className,
+  style,
+  alert,
 }) => {
   const Content = (
     <>
@@ -33,12 +47,41 @@ export const StatsCard: FunctionComponent<StatsCardProps> = ({
           centered && 'flex min-h-60 flex-col items-center justify-between',
         )}
       >
-        <CardHeader className={ctw('p-0 font-medium', centered && 'text-center')}>
+        <CardHeader
+          className={ctw(
+            'flex-row justify-between gap-4 p-0 font-medium',
+            centered && 'text-center',
+          )}
+        >
           {title}
+          {tendency && (
+            <div>
+              <div
+                className={ctw(
+                  'flex items-center gap-0.5 rounded-sm px-1 py-0 text-sm font-medium',
+                  tendency.kind === 'positive' && 'bg-green-100 text-green-600',
+                  tendency.kind === 'negative' && 'bg-red-100 text-red-600',
+                  tendency.kind === 'neutral' && 'bg-gray-100 text-gray-600',
+                )}
+              >
+                <span>{tendency.value}</span>
+                {tendency.direction === 'up' ? (
+                  <ArrowUpRightIcon className="d-4" />
+                ) : (
+                  <ArrowDownRightIcon className="d-4" />
+                )}
+              </div>
+            </div>
+          )}
         </CardHeader>
-        <p className={ctw('text-3xl font-bold', centered && 'text-center text-5xl')}>
-          {count > 0 ? `${prefix}${Intl.NumberFormat('en').format(count)}` : 0}
-        </p>
+        <div className="flex items-center gap-4">
+          {alert && <WarningFilledSvg className={'mt-1 d-10'} />}
+          <p className={ctw('text-3xl font-bold', centered && 'text-center text-5xl')}>
+            {typeof value === 'number' && value > 0
+              ? `${prefix}${Intl.NumberFormat('en').format(value)}`
+              : value ?? 0}
+          </p>
+        </div>
         <p className="text-sm text-muted-foreground">{description}</p>
       </CardContent>
     </>
@@ -46,11 +89,15 @@ export const StatsCard: FunctionComponent<StatsCardProps> = ({
 
   if (href) {
     return (
-      <Card className="shadow-md">
+      <Card className={ctw('shadow-md', className)} style={style}>
         <Link to={href}>{Content}</Link>
       </Card>
     );
   }
 
-  return <Card className="shadow-md">{Content}</Card>;
+  return (
+    <Card className={ctw('shadow-md', className)} style={style}>
+      {Content}
+    </Card>
+  );
 };