+++ packages/ui/src/components/organisms/Form/DynamicForm/hooks/internal/useClear/value-cleaners/documentfield-value-cleaner.ts
@@ -1,33 +1,57 @@
 import { AnyObject } from '@/common';
-import { request } from '@/common/hooks/useHttp';
+import { IHttpParams, request } from '@/common/hooks/useHttp';
 import { toast } from 'sonner';
-import { IDocumentFieldParams, IDocumentTemplate } from '../../../../fields';
+import {
+  getDocumentObjectFromDocumentsList,
+  IDocumentFieldParams,
+  IDocumentTemplate,
+} from '../../../../fields';
 import { getFileOrFileIdFromDocumentsList } from '../../../../fields/DocumentField/hooks/useDocumentUpload/helpers/get-file-or-fileid-from-documents-list';
 import { IFormElement, TBaseFields } from '../../../../types';
+import { DEFAULT_DELETION_PARAMS } from '../../../../fields/DocumentField/defaults';
 
 export const DOCUMENT_FIELD_VALUE_CLEANER = 'documentfield';
 
 export const documentFieldValueCleaner = async <TValue extends Array<{ id: string }>>(
   value: TValue,
   element: IFormElement<TBaseFields, IDocumentFieldParams>,
+  httpParams?: IHttpParams,
   metadata?: AnyObject,
 ): Promise<TValue | undefined> => {
   if (!Array.isArray(value)) {
     return undefined;
   }
 
-  if (element.params?.httpParams?.deleteDocument) {
-    const fileOrFileId = getFileOrFileIdFromDocumentsList(
-      value as unknown as IDocumentTemplate[],
-      element as IFormElement<'documentfield', IDocumentFieldParams>,
-    );
+  const defaultHttpParams = element.params?.httpParams?.deleteDocument || DEFAULT_DELETION_PARAMS;
 
-    if (!(fileOrFileId instanceof File)) {
-      try {
-        await request(element.params?.httpParams?.deleteDocument, metadata);
-      } catch (error) {
-        toast.error(`Failed to delete document on hide. ${(error as Error)?.message}`);
-      }
+  const mergedHttpParams = {
+    ...defaultHttpParams,
+    params: {
+      ...httpParams?.params,
+    },
+    headers: {
+      ...defaultHttpParams.headers,
+      ...httpParams?.headers,
+    },
+  };
+
+  const document = getDocumentObjectFromDocumentsList(
+    value as unknown as IDocumentTemplate[],
+    element as IFormElement<'documentfield', IDocumentFieldParams>,
+  );
+
+  const fileOrFileId = getFileOrFileIdFromDocumentsList(
+    value as unknown as IDocumentTemplate[],
+    element as IFormElement<'documentfield', IDocumentFieldParams>,
+  );
+
+  if (!(fileOrFileId instanceof File) && document?._document?.id) {
+    try {
+      await request(mergedHttpParams, metadata, {
+        ids: [document?._document?.id],
+      });
+    } catch (error) {
+      toast.error(`Failed to delete document on hide. ${(error as Error)?.message}`);
     }
   }