+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @ballerine/ui
 
+## 0.7.154
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.105
+
+## 0.7.152
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.103
+
+## 0.7.150
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.101
+
+## 0.7.149
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/common@0.9.100
+
+## 0.7.148
+
+### Patch Changes
+
+- Bump
+
 ## 0.7.147
 
 ### Patch Changes